Crate aoc_list

Implementation of lisp-like lists

A list is represented as the first list item and a reference to another list to represent the remaining items - these are referred to as the “head” and the “tail”. There is a special case for the empty list.

While this is much less efficient than a Vec, it allows reuse in a way which can be very beneficial for certain types of algorithm. Adding new items to a list only creates one new allocation per item, and removing any number of items from the start of the list requires no allocation.

We also declare CharList as List <char>, and implement some extra methods and traits to make it work nicely with String and str.

Enums

List

Type Definitions

CharList